refactor(types): move all interfaces to a dedicated file
Helps to circumvent circular dependency issues in the Rollup build.
