+++ ember-file-upload/addon/helpers/file-queue.ts
@@ -2,9 +2,9 @@ import Helper from '@ember/component/helper';
 import { registerDestructor } from '@ember/destroyable';
 import { inject as service } from '@ember/service';
 import UploadFile from '../upload-file';
-import { QueueListener } from '../queue';
 import type FileQueueService from '../services/file-queue';
 import { DEFAULT_QUEUE } from '../services/file-queue';
+import { QueueListener } from 'ember-file-upload/interfaces';
 
 interface FileQueueArgs {
   name?: string;
