+++ ember-file-upload/addon/upload-file.ts
@@ -1,72 +1,18 @@
 import { tracked } from '@glimmer/tracking';
 
-import { upload, UploadOptions } from './system/upload';
-import HTTPRequest, { HTTPRequestResponse } from './system/http-request';
+import { upload } from './system/upload';
+import HTTPRequest from './system/http-request';
 import UploadFileReader from './system/upload-file-reader';
 
 import Queue from './queue';
 import { guidFor } from '@ember/object/internals';
 import RSVP from 'rsvp';
-
-/**
- * Possible file states.
- *
- * @remarks
- *
- * Here is the statechart describing the flow of state:
- *
- * ```
- *       .------.     .---------.     .--------.
- *   o--| queued |-->| uploading |-->| uploaded |
- *       `------`     `---------`     `--------`
- *          ^              |    .-------.
- *          |              |`->| aborted |
- *          |              |    `-------`
- *          |  .------.    |    .---------.
- *          `-| failed |<-` `->| timed_out |-.
- *          |  `------`         `---------`  |
- *          `-------------------------------`
- * ```
- */
-export enum FileState {
-  Queued = 'queued',
-  Uploading = 'uploading',
-  TimedOut = 'timed_out',
-  Aborted = 'aborted',
-  Uploaded = 'uploaded',
-  Failed = 'failed',
-}
-
-export enum FileSource {
-  /**
-   * the file is created using the native file picker
-   */
-  Browse = 'browse',
-
-  /**
-   * the file was created using drag and drop from their desktop
-   */
-  DragAndDrop = 'drag-and-drop',
-
-  /**
-   * the file was created by dragging the file from another webpage
-   */
-  Web = 'web',
-
-  /**
-   * the file is created from a data URL using the `fromDataURL`
-   * method for files. This usually means that the file was created
-   * manually by the developer on behalf of the user
-   */
-  DataUrl = 'data-url',
-
-  /**
-   * the file is created from a blob using the `fromBlob`
-   * method for files. This usually means that the file was created
-   * manually by the developer
-   */
-  Blob = 'blob',
-}
+import {
+  FileSource,
+  FileState,
+  HTTPRequestResponse,
+  UploadOptions,
+} from 'ember-file-upload/interfaces';
 
 /**
  * Files provide a uniform interface for interacting
